+++ CSharpWebAdvanced-CourseProject/CinemaSystem.Web/Areas/Admin/Controllers/GenreController.cs
@@ -6,6 +6,8 @@
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using static CinemaSystem.Common.GeneralApplicationConstants;
+    using static CinemaSystem.Common.NotificationMessagesConstants;
+
 
 
     [Area(AdminArea)]
@@ -20,8 +22,16 @@
         }
         public async Task<IActionResult> Index()
         {
-            IEnumerable<GenreViewModel> genres = await genreService.GetGenresAsync();
-            return View(genres);
+            try
+            {
+                IEnumerable<GenreViewModel> genres = await genreService.GetGenresAsync();
+                return View(genres);
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = GeneralError;
+                return RedirectToAction("Dashboard", "Admin");
+            }
         }
 
         public IActionResult Add()
@@ -37,33 +47,85 @@
                 return View(genre);
             }
 
-            await genreService.AddGenreAsync(genre);
+            try
+            {
+                await genreService.AddGenreAsync(genre);
+
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = GeneralError;
+            }
             return RedirectToAction(nameof(Index));
         }
 
         public async Task<IActionResult> Edit(string id)
         {
-            GenreAddEditViewModel? model = await genreService.GetEditGenreModelAsync(id);
-
-            return View(model);
+            if (id == null)
+            {
+                TempData[ErrorMessage] = GeneralError;
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                GenreAddEditViewModel? model = await genreService.GetEditGenreModelAsync(id);
+                if (model != null)
+                {
+                    return View(model);
+                }
+                else
+                {
+                    TempData[ErrorMessage] = GeneralError;
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = GeneralError;
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         [HttpPost]
         public async Task<IActionResult> Edit(string id, GenreAddEditViewModel genre)
         {
+            if (id == null)
+            {
+                TempData[ErrorMessage] = GeneralError;
+                return RedirectToAction(nameof(Index));
+            }
             if (!ModelState.IsValid)
             {
                 return View(genre);
             }
 
-            await genreService.EditGenreAsync(id, genre);
+            try
+            {
+                await genreService.EditGenreAsync(id, genre);
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = GeneralError;
+            }
             return RedirectToAction(nameof(Index));
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(string id)
         {
-            await genreService.DeleteGenreAsync(id);
+            if (id == null)
+            {
+                TempData[ErrorMessage] = GeneralError;
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                await genreService.DeleteGenreAsync(id);
+            }
+            catch (Exception)
+            {
+                TempData[ErrorMessage] = GeneralError;
+            }
             return RedirectToAction(nameof(Index));
         }
     }
